Descriptors
The descriptor set and layout is now created based on the shader
reflection.  The app still kind of assumes you have a full UBO and
possibly a sampler, but this will change soon.  You can now change the
binding index and things will still work, which didn't happen before.

The next step is to dynamically build the UBO.  Then dynamically bind
the samplers.  Finally, when that works well, add texture support for
models and custom textures such as noise maps.