Audio/Analysis Framework
Ported from earlier work.  This is work in progress support for
audio->texture support.

Effectively enabling the graphics to change based on the sound
frequencies/beat.

There is an audio device, with a UI for configuring input/output
channels, and an FFT analysis framework for converting the audio wave to
audio spectrum buckets.

The analysis code is designed to efficiently run in the audio thread (or
even a seperate thread - though the work is fast enough that this isn't
required currently).  There is a locking mechanism to access the audio
data in the UI without ever blocking the audio thread or forcing it to
lock a mutex (which is not advised for audio threads because it can
cause the thread to lose its time slice and threaten the integrity of
the delivered audio).

This code is compiling but not currently called.  It should be up and
running soon.