Audio Configuration
The audio dialog now works and saves/restores state.
There is a bit more work to do to synchronize the changes and ensure
everything is updated with the new settings; which can cause crashes
when changing some things.  Should be fixed soon.
Audio input support is very nearly as far as I want to take it at this
point.  I need to add the vec4 spectrum band input as a shader variable
and clean up a few things here and there.