Audio
Moved audio processing onto a thread.
This is also prep for cleaner reset of the audio when the user changes
the config.

When the audio ticks and input (and in future output) audio is
generated, it is pushed onto an audio channel queue (which is lock free).
The analysis thread can then pull audio off the queue and process it in a seperate
thread.