Pass Rework
A cleaner setup of pass state.  I still have more to do, but this is
stable again.  The operations required to update the pass are cleaner
now and separated into steps - 'prepare_rendertargets',
'prepare_bindings', etc.  I will likely refine these further.  I'm
thinking more of a dynamic setup here, where bindings and things are
evaluated as needed and then thrown away; with options for caching.

The other thing that this CL does is support a ping-pong approach to
backbuffers, where you can always _read_ the contents of the previous
target that you are writing.  Shadertoy does this alot, and in a
transparent way - so you just sample and write the target at the same
time, even though under the covers it is managing a pair of buffers for
you.

I'm happier with this cleaner approach, anyway.  Though there is still a
way to go.