Deferred Rendering/Multi Targets
A demo of deferred rendering is in the works, and will likely need more
tweaks, but this is at a good place and has other fixes, so I'm merging
it to main.

The new camera support still requires the option to move the camera
using mouse/keyboard, and a reset option to reset it to the default
value.

The next step is to add a quad pass and 'draw' a light into the scene.
Then I'll be thinking about how to at least crudely specify lights in
the scene and maybe even move them.

- Can now render to more than one Targets.
- Can now specify a camera and its properties and use it in a pass.
  default_camera is created and used by default if not specified
- Added a float format surface too.