Model texture WIP, and updates to latest zep/vcpkg
It is my intention to do a full glb-loading PBR support.  For that, I need to load textures, including HDR ones, and apply them to models,
as well as supply the descriptor data to shaders.  I've run out of steam
for now, so checking this one in until I get to it.

Zep was also recently updated, and vcpkg.