Refactoring the window
It's always been my intent to keep the vulkan device seperate; but I
haven't managed it everywhere.  This is just a little cleanup to get
closer to that goal, and make the rendering seperate to the display of
the window.  I'll keep refining it.

Also added a show state for the Targets window in the menu.